tests/performance: reconcile until idle before benchmark
We'd like to run benchmarks starting from a steady state.
To this end, do a reconciliation round before proceeding with the
benchmark.

This is useful for benchmarks that use tenant dir snapshots since
a non-standard tenant configuration is used to generate the snapshot.
The storage controller is not aware of the non default tenant configuration
and will reconcile while the bench is running.