Misc cleanup of #includes and comments in the neon extension
Remove useless and often wrong IDENTIFICATION comments. PostgreSQL
sources have them, mostly for historical reasons, but there's no
need for us to copy that style.

Remove unnecessary #includes in header files, putting the #includes
directly in the .c files that need them. The principle is that a
header file should #include other header files if they need
definitions from them, such that each header file can be compiled on
its own, but not other #includes. (There are tools to enforce that,
but this was just a manual clean up of violations that I happened to
spot.)