pageserver: suppport scattered queries
The core of the change is in `Timeline::get_vectored_init_fringe`.
When preparing the fringe, we push the next layers for each keyspace
into it.

Rest of the changes should be self explanatory.

Important Note: We only support scattered queries for disjoint
key spaces. In other words, requesting the same key at different
LSNs is not supported. A runtime assertion in
`VersionedKeySpaceQuery::total_keyspace` will fire if this invariant
is violated. It will be enforced by the page service batching code.