pageserver: do not coalesce reads ending at different LSNs
Previously the fringe was keyed by layer and start LSN of the
read. This meant that a read within the same layer with a different
end LSN, would be coalesced in the existing read incorrectly.

This situation is not possible for uniform reads, but very much
so for scattered reads.

The fix is very simple: add the end LSN of the read to the fringe key.
In practical terms, we can end up visiting a layer more than once.
For uniform reads this shouldn't happen - I expect the number of layer
visits to stay the same. For scattered reads, I expect it to happen
within the max delta of the provided LSN ranges.

An alternative solution would be to bake this logic into the layer
reading logic itself and teach it how to read scattered keyspaces.
That's intrusive and error prone. Indices are mostly cached in memory
so it shouldn't be too costly to read it multiple times. Hence, I
prefered the simple fringe approach.