refactor: Extract some code in pagebench getpage command to function
This makes it easier to add a different client implementation
alongside the current one. I started working on a new gRPC-based
protocol to replace the libpq protocol, which will introduce a new
function like `client_libpq`, but for the new protocol.

It's a little more readable with less indentation anyway.