Make all SSL parameters for walproposer connections configurable
In preparation for TLS enablement of the walproposer -> safekeeper
connection, add GUCs to control the various SSL connection parameters.
Start with the current values that we use (Postgres defaults), and
change them when we have all the files in place for enabling TLS
encryption on the connections.

Part-of: https://github.com/neondatabase/cloud/issues/25823
Signed-off-by: Tristan Partin <tristan@neon.tech>