pageserver: handle empty get vectored queries
Problem

If all batched requests are excluded from the query by
`Timeine::get_rel_page_at_lsn_batched` (e.g. because they
are past the end of the relation), the read path would panic
since it doesn't expect empty queries. This is a change in behaviour
that was introduced with the scattered query implementation.

Summary of Changes

Handle empty queries explicitly.