storage_controller: don't kill the process on long step down
The idea was to avoid the split brain scenario if split down
takes a long time and the new pod proceeds as leader in the meantime
by killing the old pod after a fixed time-out.

This does not work in production where we use k8s. The pod will simply
be restarted and attempt to become leader again. When it succeeds,
k8s will kill it since it still belongs to the old replica set.

Obviously, this behaviour is not desirable, so we let step down
finish and log a warning every 10 seconds if it's taking long to do so.